refactor: 타입 일관성 및 코드 정리
- WhiteNoiseItem을 별도 DTO 파일로 추출 (BlockedAppItem과 동일 패턴)
- Entity 필드 Integer → Long 통일 (durationMinutes, totalMinutes)
- calculateTotalMinutes()에서 mapToLong 사용
- getMonthlyRecords에 memberId 파라미터 추가
- ConstraintViolationException에서 propertyPath 로그로만 출력

+++ src/main/java/whoreads/backend/domain/readingsession/entity/ReadingInterval.java
@@ -30,7 +30,7 @@ public class ReadingInterval {
     private LocalDateTime endTime;
 
     @Column(name = "duration_minutes")
-    private Integer durationMinutes;
+    private Long durationMinutes;
 
     @Builder
     public ReadingInterval(ReadingSession readingSession, LocalDateTime startTime) {
@@ -45,7 +45,7 @@ public class ReadingInterval {
         if (endTime.isBefore(this.startTime)) {
             throw new IllegalArgumentException("종료 시간은 시작 시간 이후여야 합니다.");
         }
-        this.durationMinutes = (int) ChronoUnit.MINUTES.between(this.startTime, endTime);
+        this.durationMinutes = ChronoUnit.MINUTES.between(this.startTime, endTime);
         this.endTime = endTime;
     }
